requirements.txt
black==24.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==1.6.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
gym==0.21.0
gym-notices==0.0.8
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.5.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.19.5
opencv-python==4.5.5.64
packaging==24.2
pathspec==0.12.1
Pillow==8.4.0
platformdirs==4.3.6
protobuf==3.20.1
pybullet==3.2.5
pyclean==3.1.0
pyglet==1.4.11
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tensorboardX==2.5.1
tomli==2.2.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
